Correct BOOLS

+++ Source/scrollrt.cpp
@@ -887,7 +887,7 @@ void scrollrt_draw_clipped_dungeon(BYTE *pBuff, int sx, int sy, int dx, int dy, 
 		}
 	}
 	if (bArch != 0) {
-		cel_transparency_active = (BYTE)TransList[bMap];
+		cel_transparency_active = TransList[bMap];
 		Cel2DecodeLightTrans(pBuff, pSpecialCels, bArch, 64, 0, 8);
 	}
 }
@@ -1353,7 +1353,7 @@ void scrollrt_draw_clipped_dungeon_2(BYTE *pBuff, int sx, int sy, int skipChunks
 		}
 	}
 	if (bArch != 0) {
-		cel_transparency_active = (BYTE)TransList[bMap];
+		cel_transparency_active = TransList[bMap];
 		Cel2DecodeLightTrans(pBuff, pSpecialCels, bArch, 64, CelSkip, 8);
 	}
 }
@@ -1773,7 +1773,7 @@ void scrollrt_draw_dungeon(BYTE *pBuff, int sx, int sy, int capChunks, int CelCa
 		}
 	}
 	if (bArch != 0) {
-		cel_transparency_active = (BYTE)TransList[bMap];
+		cel_transparency_active = TransList[bMap];
 		CelDecodeHdrLightTrans(pBuff, pSpecialCels, bArch, 64, 0, CelCap);
 	}
 }
